Create title screen fragment by 	Creating new kotlin resource file for title 	DON'T create fragment_title.xml because it already exists
Going into activity_main.xml and adding the new fragment

Within a fragment, setContentView does not exist. However, the project already implements data binding.
So to grab the FragmentTitle we run DataBindingUtil.inflate and return binding.root. We want to return
the root of the layout, not the inflated layout .

Enter: The NavHostFragment. The NavHostFragment is like a waiter attending the buffet line, he is
a Person (Fragment) within the Activity (Restaurant), but his job is not to eat food (display Views).
His job is to manage the buffet line (the backstack). Then, the customers (other Fragments) make it
their business to display content.

1) Create a NavHostFragment using New resource type->Android resource type. This navigation.xml
file will land in res/navigation separate from the other resource files.
2) In activity_main.xml we replace the TitleFragment within the fragment XML with our new navhostfragment
3) Set it as the default nav host to intercept the back key, give it an id, and tell the fragment xml
which nav resource to use.
4) Add a fragment to the navigation graph with the phone plus button. Then add the game fragment and connect
the two.
5) Finally, by adding an onclick listener from the title fragment we can programatically travel from
one fragment to the next. This is done by:
	Going back to TitleFragment.kt
	find the playButton (class attribute of binding) and have it run setOnClickListener on...
	Navigation.createNavigateOnClickListener(R.id.action_titleFragment_to_gameFragment)
This utilizes androidktx